requirements.txt
django==3.1.7
factory-boy==3.2.0
pytest-cov==2.11.1
pytest-django==4.1.0
pytest-flake8==1.0.7
freezegun==1.1.0